Created file structure for refactor
